+++ persistence/src/main/java/de/starwit/persistence/entity/ExecutionPolicies.java
@@ -1,6 +1,0 @@
-package de.starwit.persistence.entity;
-
-public enum ExecutionPolicies {
-    MANUAL, WITHCHECK, AUTOMATIC;
-
-}