+++ persistence/src/main/resources/db/migration/V1_2__add_action_state.sql
@@ -1,0 +1,8 @@
+ALTER TABLE "actiontype"
+ADD "endpoint" VARCHAR(255);
+
+ALTER TABLE "action"
+ADD "state" VARCHAR(100);
+
+ALTER TABLE "action"
+ADD "metadata" VARCHAR(500);